requirements.txt
# Plant Disease Detection System — Requirements
# Install: pip install -r requirements.txt

# Core ML
tensorflow>=2.12.0
numpy>=1.23.0
scikit-learn>=1.2.0

# Image processing
opencv-python-headless>=4.7.0
Pillow>=9.4.0

# Web App
streamlit>=1.28.0

# Data & Visualization
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Utilities
gdown>=4.7.0       # model download from Google Drive
